requirements.txt
python-documentcloud==4.1.3
requests
happytransformer
nltk
